+++ src/coral_fans/commands/Calculate.cpp
@@ -1,4 +1,5 @@
 #include "coral_fans/functions/calculate/Calculate.h"
+#include "coral_fans/Config.h"
 #include "coral_fans/base/Macros.h"
 #include "ll/api/command/CommandHandle.h"
 #include "ll/api/command/CommandRegistrar.h"
@@ -7,17 +8,19 @@
 #include "ll/api/i18n/I18n.h"
 #include "mc/server/commands/CommandOrigin.h"
 #include "mc/server/commands/CommandOutput.h"
-#include "mc/server/commands/CommandPermissionLevel.h"
 
 
 namespace coral_fans::commands {
 
-void registerCalculateCommand(CommandPermissionLevel permission) {
+void registerCalculateCommand(config::CommandConfigStruct& config) {
     using ll::i18n_literals::operator""_tr;
 
     // reg cmd
-    auto& calculateCommand = ll::command::CommandRegistrar::getInstance(false)
-                                 .getOrCreateCommand("calculate", "command.calculate.description"_tr(), permission);
+    auto& calculateCommand = ll::command::CommandRegistrar::getInstance(false).getOrCreateCommand(
+        config.command,
+        "command.calculate.description"_tr(),
+        config.permission
+    );
     calculateCommand.runtimeOverload().text("pt").execute(
         [](CommandOrigin const& origin, CommandOutput& output, ll::command::RuntimeCommand const&) {
             COMMAND_CHECK_PLAYER