+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.9.2] - 2026-4-14
+
+### Added
+
++ Added locate duplicatable feature 
++ Added minerule fuck_population_cap
++ Added more config options
+
 ## [2.9.2] - 2026-3-1
 
 ### Added